hmicontroller: Pass INVALID_ID to auto-generate layer ID
A layer ID of zero is not correct.  I am guessing that the previous
intention was to use zero to generate an automatic ID, and the ILM
documentation incorrectly said that it would (upstream bug [LM-6]).
INVALID_ID should be passed instead, to get a generated ID.

The parameter to helper function createLayer() must now be passed by
reference so that the member variable m_backgroundSurfaceId can be
assigned the right value. The layer ID used to be constant (zero) and
therefore this was not the case before.

[GDP-778] wayland-ivi-extension layer id integration issues from P-0.1
          and P-1.0 upgrade

[LM-6] Documentation for ilm_layerCreateWithDimension() is wrong
       regarding creating new ID automatically?

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>