gdp-hmi: Avoid error Layer 0 already exists
This is a temporary fix to the problem of getting an error in
layer management interaction because layer 0 being created
more than once.

When LayerManager version was updated it was more strict on some input
and uncovered bugs that were likely there, but just happened to work
before. Also, the LM docs were wrong and had stated that a layer ID of
zero could be passed in order for a dynamic one to be created by the
framework. This is not true (bug has been filed/fixed).

But it's possible this led original authors to use ID 0 as input
to create layer.

It's also possible that the bug remains - that there is some reason that
the value zero pops up and is being used. This patch only forces the
background ID to another non-zero value which seems to avoid the
problem.

If you notice that the background _layer ID_ is created with the same ID
as the (also hard coded) background _surface ID_. This was done that way
also before.  When no longer using layer ID zero, the layer was no
longer created before first use.  To counteract, the creation of the
background layer is now guaranteed at early startup.

This code needs a bigger overhaul - it's for now just some quick fixes
to the symptoms.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>